Refactor node statistics and counters from `int` to `long` to support larger datasets, update related tests, visitors, and builders accordingly.

+++ search/src/main/java/net/chesstango/search/smart/features/statistics/node/filters/AlphaBetaStatisticsVisited.java
@@ -13,7 +13,7 @@ import net.chesstango.search.smart.alphabeta.filters.AlphaBetaFilter;
 public class AlphaBetaStatisticsVisited implements AlphaBetaFilter {
     private AlphaBetaFilter next;
 
-    private int[] visitedNodesCounters;
+    private long[] visitedNodesCounters;
 
     @Override
     public void accept(Visitor visitor) {
